refactor(nznuxsir): update code
