+++ components/features/comment/modal-see-more/index.tsx
@@ -3,7 +3,7 @@
 import React from "react";
 import isEqual from "react-fast-compare";
 
-import AppLink from "@/components/common/AppLink";
+import AppLink from "@/components/common/app-link";
 import { Button } from "@/components/ui/button";
 import {
   DialogClose,
