+++ theme/index.ts
@@ -5,6 +5,6 @@ import resolveConfig from "tailwindcss/resolveConfig";
 
 import tailwindConfig from "../tailwind.config";
 
-const tailwind: Config = resolveConfig(tailwindConfig) as Config;
+const tailwind: Config = resolveConfig(tailwindConfig) as unknown as Config;
 
 export default tailwind;
